# rfid-backend-deploy/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RFID Attendance API",
    "version": "1.0.0",
    "description": "VIPSTC RFID-Based Attendance System API"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AuthReq": {
        "type": "object",
        "required": ["uid"],
        "properties": { "uid": { "type": "string" } }
      },
...
      "ScanReq": {
        "type": "object",
        "required": ["uid","sessionId"],
        "properties": {
          "uid":       { "type": "string" },
          "sessionId": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/device/auth": {
      "post": {
        "summary": "Authenticate device with Teacher RFID",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthReq" }
            }
          }
        },
        "responses": {
          "200": { "description": "Authenticated" }
        }
      }
    },
...
    "/api/scan": {
      "post": {
        "summary": "Record a scan",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ScanReq" }
            }
          }
        },
        "responses": {
          "204": { "description": "Recorded" }
        }
      }
    },
...
  }
}